API

Overview

In order to support InteractivePrompts, you need to specify how a certain display should handle (see Interface functions) a certain question (see Question types).

While this page does detail the requirements, it is probably best paired with an example to get a sense of how this might look in practice.

Question types

Foundational types

InteractivePrompts.Prompt — Type

abstract type Prompt{T} end

The supertype of all specific prompt forms. Should produce an Answer{T}.

TODO improve docstring

InteractivePrompts.Behaviour — Type

struct Behaviour

The particular way a question should behave. This may consist of

A default value (when applicable)

A placeholder value (when applicable)

A timeout applied to the question (in seconds)

A callback function

Whether the question should "trap" focus (trapfocus)

Whether the question should be able to be aborted (abotable)

TODO improve docstring

InteractivePrompts.Style — Type

struct Style

A small selection of basic styling attributes that should be easily represented across multiple display types.

These attributes are:

Whether the question should be shown in a compact form

The border the question should (or should't) have

The style of the question's text

Whether the question should be shown in a modal format

TODO improve docstring

InteractivePrompts.Question — Type

struct Question{P <: Prompt}

A question, ready to be asked, consisting of:

The prompt (p) that the user answers

The behaviour the question should exhibit

The style the question should be shown with

TODO improve docstring

InteractivePrompts.Answer — Type

struct Answer{T}

An answer from a Question{Prompt{T}}. Holds a response that is either of type T, or the value nothing.

The flag aborted indicates whether the question was aborted or not.

TODO improve docstring

Specific forms

InteractivePrompts.Confirmation — Type

struct Confirmation <: Menu{Bool}

A yes/no confirmation, optionally with custom yes/no strings.

TODO improve docstring

Interface functions

InteractivePrompts.render — Function

render(d::AbstractDisplay, q::Question)

Render the question q to the display d.

Produces a representaton of the question's state.

TODO improve docstring

InteractivePrompts.clear — Function

clear(d::AbstractDisplay, q::Question, state)

Clear the previous rendering of q in the display d.

TODO improve docstring

InteractivePrompts.abort — Function

abort(d::AbstractDisplay, [q::Question, state::Any])

Abort the attempt to answer q in the display d.

Methods accepting q and state arguments must return the new state.

TODO improve docstring

InteractivePrompts.focus — Function

focus(d::AbstractDisplay, [q::Question, state::Any])

Give focus to the question q in display d. Exactly one question should have focus at any one time.

Methods accepting q and state arguments must return the new state.

TODO improve docstring

InteractivePrompts.unfocus — Function

unfocus(d::AbstractDisplay, [q::Question, state::Any])

Remove focus from the question q in display d. Exactly one question should have focus at any one time.

Methods accepting q and state arguments must return the new state.

TODO improve docstring

InteractivePrompts.answer — Function

answer(d::AbstractDisplay, q::Question, state)

Extract the answer to q (show in display d) from state.

TODO improve docstring